get query for subscriptions
