+++ apps/consultation-portal/screens/Subscriptions/getSubscriptionCases.ts
@@ -1,0 +1,15 @@
+import gql from 'graphql-tag'
+
+export const GET_SUBSCRIPTION_CASES = gql`
+  query consultationPortalAllCases {
+    consultationPortalAllCases {
+      id
+      caseNumber
+      name
+      institutionName
+      policyAreaName
+    }
+  }
+`
+
+export const GET_TYPES = gql``
