+++ apps/consultation-portal/types/interfaces.ts
@@ -4,15 +4,15 @@ export interface Case {
   id: number
   caseNumber: string
   name: string
-  adviceCount: number
-  shortDescription: string
-  statusName: string
-  institutionName: string
-  typeName: string
-  policyAreaName: string
-  processBegins: string
-  processEnds: string
-  created: string
+  adviceCount?: number
+  shortDescription?: string
+  statusName?: string
+  institutionName?: string
+  typeName?: string
+  policyAreaName?: string
+  processBegins?: string
+  processEnds?: string
+  created?: string
 }
 
 export interface ArrOfIdAndName {
